requirements.txt
pandas==1.5.3
numpy==1.24.2
finta
matplotlib==3.7.1
scipy==1.10.1
requests==2.28.2
websocket-client==1.5.1
statsmodels==0.13.5
